Define and use macro to create context wrapper in nightly Rust.
