+++ src/context.rs
@@ -30,7 +30,7 @@ pub trait TransparentContextMut<T> {
 // so we would get into trouble with specialization.
 #[cfg(feature = "stable")]
 #[macro_export]
-macro_rules! wrap_context{
+macro_rules! wrap_context {
     ($type_name:ty, $wrapper_name:ident) => {
         pub struct $wrapper_name<'a, 'c, C> {
             aspect: &'a mut $type_name,
@@ -68,5 +68,101 @@ macro_rules! wrap_context{
                 self.child_context.get_mut()
             }
         }
-    }
+    };
+}
+
+#[cfg(not(feature = "stable"))]
+macro_rules! wrap_context {
+    ($type_name:ty, $wrapper_name:ident) => {
+        pub struct $wrapper_name<'a, 'c, C> {
+            aspect: &'a mut $type_name,
+            child_context: &'c mut C,
+        }
+
+        // Thanks to Lymia for this trick.
+        // For more info, see
+        // https://github.com/rust-lang/rust/issues/31844#issuecomment-397650553
+        trait UniversalTransparentContext<T> {
+            fn get(&self) -> &T;
+        }
+
+        trait UniversalTransparentContextMut<T> {
+            fn get_mut(&mut self) -> &mut T;
+        }
+
+        impl<'a, 'c, C, T> UniversalTransparentContext<T> for $wrapper_name<'a, 'c, C> {
+            default fn get(&self) -> &T {
+                unreachable!();
+            }
+        }
+
+        impl<'a, 'c, C, T> UniversalTransparentContext<T> for $wrapper_name<'a, 'c, C>
+        where
+            C: TransparentContext<T>,
+        {
+            fn get(&self) -> &T {
+                self.child_context.get()
+            }
+        }
+
+        impl<'a, 'c, C, T> UniversalTransparentContextMut<T> for $wrapper_name<'a, 'c, C> {
+            default fn get_mut(&mut self) -> &mut T {
+                unreachable!();
+            }
+        }
+
+        impl<'a, 'c, C, T> UniversalTransparentContextMut<T> for $wrapper_name<'a, 'c, C>
+        where
+            C: TransparentContextMut<T>,
+        {
+            fn get_mut(&mut self) -> &mut T {
+                self.child_context.get_mut()
+            }
+        }
+
+        trait GenericOrSpecial<T> {}
+
+        impl<'a, 'c, C, T> GenericOrSpecial<T> for $wrapper_name<'a, 'c, C> where
+            C: TransparentContext<T>
+        {
+        }
+
+        impl<'a, 'c, C> GenericOrSpecial<$type_name> for $wrapper_name<'a, 'c, C> {}
+
+        impl<'a, 'c, C, T> TransparentContext<T> for $wrapper_name<'a, 'c, C>
+        where
+            $wrapper_name<'a, 'c, C>: GenericOrSpecial<T>,
+        {
+            default fn get(&self) -> &T {
+                <Self as UniversalTransparentContext<T>>::get(self)
+            }
+        }
+
+        impl<'a, 'c, C> TransparentContext<$type_name> for $wrapper_name<'a, 'c, C>
+        where
+            $wrapper_name<'a, 'c, C>: GenericOrSpecial<$type_name>,
+        {
+            fn get(&self) -> &FrameCounter {
+                self.aspect
+            }
+        }
+
+        impl<'a, 'c, C, T> TransparentContextMut<T> for $wrapper_name<'a, 'c, C>
+        where
+            $wrapper_name<'a, 'c, C>: GenericOrSpecial<T>,
+        {
+            default fn get_mut(&mut self) -> &mut T {
+                <Self as UniversalTransparentContextMut<T>>::get_mut(self)
+            }
+        }
+
+        impl<'a, 'c, C> TransparentContextMut<$type_name> for $wrapper_name<'a, 'c, C>
+        where
+            $wrapper_name<'a, 'c, C>: GenericOrSpecial<$type_name>,
+        {
+            fn get_mut(&mut self) -> &mut $type_name {
+                self.aspect
+            }
+        }
+    };
 }
