+++ src/middleware/frame_counter.rs
@@ -18,107 +18,7 @@ impl FrameCounter {
     }
 }
 
-#[cfg(not(feature = "stable"))]
-pub struct FrameCounterContext<'sc, 'cc, C> {
-    aspect: &'sc mut FrameCounter,
-    child_context: &'cc mut C,
-}
-
-#[cfg(feature = "stable")]
 wrap_context!(FrameCounter, FrameCounterContext);
-
-#[cfg(not(feature = "stable"))]
-pub mod nightly {
-    use super::{FrameCounter, FrameCounterContext};
-    use crate::context::{TransparentContext, TransparentContextMut};
-
-    // Thanks to Lymia for this trick.
-    // For more info, see
-    // https://github.com/rust-lang/rust/issues/31844#issuecomment-397650553
-    trait UniversalTransparentContext<T> {
-        fn get(&self) -> &T;
-    }
-
-    trait UniversalTransparentContextMut<T> {
-        fn get_mut(&mut self) -> &mut T;
-    }
-
-    impl<'sc, 'cc, C, T> UniversalTransparentContext<T> for FrameCounterContext<'sc, 'cc, C> {
-        default fn get(&self) -> &T {
-            unreachable!();
-        }
-    }
-
-    impl<'sc, 'cc, C, T> UniversalTransparentContext<T> for FrameCounterContext<'sc, 'cc, C>
-    where
-        C: TransparentContext<T>,
-    {
-        fn get(&self) -> &T {
-            self.child_context.get()
-        }
-    }
-
-    impl<'sc, 'cc, C, T> UniversalTransparentContextMut<T> for FrameCounterContext<'sc, 'cc, C> {
-        default fn get_mut(&mut self) -> &mut T {
-            unreachable!();
-        }
-    }
-
-    impl<'sc, 'cc, C, T> UniversalTransparentContextMut<T> for FrameCounterContext<'sc, 'cc, C>
-    where
-        C: TransparentContextMut<T>,
-    {
-        fn get_mut(&mut self) -> &mut T {
-            self.child_context.get_mut()
-        }
-    }
-
-    trait GenericOrSpecial<T> {}
-
-    impl<'sc, 'cc, C, T> GenericOrSpecial<T> for FrameCounterContext<'sc, 'cc, C>
-    where
-        C: TransparentContext<T>
-    {
-    }
-
-    impl<'sc, 'cc, C> GenericOrSpecial<FrameCounter> for FrameCounterContext<'sc, 'cc, C> {}
-
-    impl<'sc, 'cc, C, T> TransparentContext<T> for FrameCounterContext<'sc, 'cc, C>
-    where
-        FrameCounterContext<'sc, 'cc, C>: GenericOrSpecial<T>,
-    {
-        default fn get(&self) -> &T {
-            <Self as UniversalTransparentContext<T>>::get(self)
-        }
-    }
-
-    impl<'sc, 'cc, C> TransparentContext<FrameCounter> for FrameCounterContext<'sc, 'cc, C>
-    where
-        FrameCounterContext<'sc, 'cc, C>: GenericOrSpecial<FrameCounter>,
-    {
-        fn get(&self) -> &FrameCounter {
-            self.aspect
-        }
-    }
-
-    impl<'sc, 'cc, C, T> TransparentContextMut<T> for FrameCounterContext<'sc, 'cc, C>
-    where
-        FrameCounterContext<'sc, 'cc, C>: GenericOrSpecial<T>,
-    {
-        default fn get_mut(&mut self) -> &mut T {
-            <Self as UniversalTransparentContextMut<T>>::get_mut(self)
-        }
-    }
-
-    impl<'sc, 'cc, C> TransparentContextMut<FrameCounter> for FrameCounterContext<'sc, 'cc, C>
-    where
-        FrameCounterContext<'sc, 'cc, C>: GenericOrSpecial<FrameCounter>,
-    {
-        fn get_mut(&mut self) -> &mut FrameCounter {
-            self.aspect
-        }
-    }
-}
 
 pub trait WithFrameCounter {
     fn frame_counter(&self) -> &FrameCounter;
